requirements.txt
joblib==1.4.2
keras==3.5.0
matplotlib==3.10.3
numpy==2.2.6
pandas==2.2.3
plotly==6.0.0
pywin32==306
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
streamlit==1.43.0
tensorflow==2.16.0rc0
tensorflow_intel==2.16.0rc0
xgboost==3.0.2